requirements.txt
openai==0.28
streamlit
pandas
xlsxwriter
python-dotenv
openpyxl
requests
schedule
APScheduler
plotly